avatar: Adjust types to acknowledge bots can be missing avatar_url.
Anywhere we're referring to an `avatarUrl` and already permit it to
be `null`, we're presumably just pulling it off of a `User` object...
which almost certainly is really a `UserOrBot`.  So we should be
prepared for the value to also be `undefined`, as that's what we'd
get on a cross-realm bot from a pre-1.9 server.

Mostly a mechanical edit:

  $ perl -i -0pe '
      s/avatarUrl: \Kstring \| null/?string/g
    ' src/**/*.js

That then cleared up two fixmes, raising warnings because they became
unused; and in both cases the code got to be a bit simplified now that
there was no longer a need to isolate the fixme-requiring step.

Also add a short TODO note about how it'd be preferable to handle this
when we eventually add a bit of a "crunchy shell" to the handling of
initial data on register.