oauth: Fold together split ios/android files.
One of the few remaining examples of #3018.

Note that it's definitely fine to import unconditionally from
`react-native-safari-view` on Android... because our `utils/openLink`
already does, and in fact does so in order to provide the exact same
functionality as one of these functions. :-)  So this illustrates
again how the .ios.js / .android.js split can hide opportunities to
share logic.