fetch: Remove faulty logic in needFetchAtFirstUnread.
Fixes #3441.

Remove the incorrect part of the `needFetchAtFirstUnread` function.
The remaining code is a small subset of the intended behavior but
is enough to fix the bad behavior.  In further commits we will make
this more sophisticated.

The initial behavior of this function was meant to mean: 'Are we
far enough from the beginning and the end of the message list to
not need to fetch more messages?'.  It is not clear if this ever
worked very well (maybe not) but right now the code is not even
trying to implement that behavior.