connect: Mass-migrate to our type-fixing wrapper.
This converts every reference to `connect` in our codebase to
go through this wrapper, giving us an easy single point of
control for the types Flow sees for `connect`.

If in the future we start using other features of `connect`
that the wrapper ignores, we can add them there; or add a
second wrapper with a different name, if that happens to be
easier.  (No reason we have to rely as heavily on overloading
as the upstream `connect` does.)

Done mechanically, like so:

  $ perl -i -0pe '
      s<^import \{ connect \} from .react-redux..\n>
       <>m || next;
      s<from .[./]*/types.;\n\K>
       <import { connect } from '\''../react-redux'\'';\n>ms;
    ' src/**/*.js