style: Mostly switch to specific types ViewStyleProp, TextStyleProp.
Some change in the upgrade to RN v0.59 (or perhaps to Flow v0.92
which accompanies it) causes Flow to start giving errors whenever we
pass a value of the type `DangerouslyImpreciseStyleProp` -- which is
what our `Style` is actually an alias for -- to a component like
`View` or `TextInput`.

So, just fix most of them to the specific style type they ought
to be.  Turns out this isn't hard; mostly just follow the error
messages where they lead.