webview js: Add `touchcancel` listener.
This event fires when a touch point "has been disrupted in some way",
according to docs on MDN; for example:

 1) Some event, like a modal alert, which might occur during
    the touch event cycle.

 2) The touch pointer goes out of focus of the current viewport.

 3) The multitouch support can't handle the amount of touches
    currently on the screen.

See https://developer.mozilla.org/en-US/docs/Web/API/TouchEvent .

It's a close cousin of `touchend` and `touchmove`, and generally
should be handled in a similar way.

This addresses a small bit of #3439.