kotlin: Drop a spurious exception-check, yay.
This exception can't be caused here by mere unexpected data from the
outside.  Instead, it's the kind of exception that would reflect a
programming error.

Specifically: if this exception is ever thrown, that's a bug in this
method.  The method starts from a proper `URL` object, and just
applies a relative URL reference which it itself constructs.

In general, it doesn't make sense to constantly try in the midst of
application code to catch exceptions that would mean a programming
error -- it just makes things more complicated, harder to reason
about, and more brittle.  The only reason we previously had such a
check was that Java's "checked exceptions" forced us to.

Happily, Kotlin abolishes "checked exceptions"!
  https://kotlinlang.org/docs/reference/exceptions.html
So we can just drop this counterproductive `catch`.

The calling code gets a bit simpler, too, by not having to worry about
getting `null`.