flow: Put in jsdoc some helpful info from recent commits.
This is information that was in the commit messages of the commits
that added these bits of code.  It's useful information, and important
not only for understanding the change (i.e. for comparing old and new
versions of the code), but instead mainly for understanding the new
code as it is.  So, the best place for it is in comments on the code
itself, right in the tree.