tests: Make a start at some well-typed example data for tests.
This will help us to write unit tests that deal with large objects
like users and messages
 * while passing genuine well-typed data,
 * and with less repetition of boilerplate boring bits of data.

This contains enough to write the particular tests I wanted to write
today, but it's quite incomplete -- there are plenty more types of
objects, and also directions in which to customize the data, that we
might add to it for the sake of other tests.

Also, as the jsdoc comments announce, some of these property values
may not look anything like they would in a real user/message/etc.!
They match our types, and I've made them plausible values where I know
off the top of my head what values are plausible, but e.g. `User` has
`timezone: string`, which isn't very informative and I don't know what
those strings are actually supposed to be.  We can fix such details as
we run into them -- fortunately, because this is test code, we'll
always know it when we do.