android: Upgrade Gradle to 5.1.1, and Android Gradle Plugin to 3.4.1.
Plugin release notes here:
  https://developer.android.com/studio/releases/gradle-plugin#3-4-0
Nothing super exciting; main thing is it requires Gradle 5.1.1.

Gradle release notes here:
  https://docs.gradle.org/5.1.1/release-notes.html
but mostly here, for upgrading from 4.x as we are:
  https://docs.gradle.org/5.1.1/userguide/upgrading_version_4.html

Probably the most interesting new feature: Gradle config files
(like `build.gradle`) can now be written in Kotlin, instead of in
Groovy as they are now.  This apparently comes with much better IDE
support too, which makes sense as Kotlin has static types (whee!).

Also a small change I noticed which hopefully is helpful: apparently
before 5.0 the `gradle` command would by default allocate a whole
GiB of RAM, just for the CLI?!  Yikes.  Now it defaults to 64 MiB.
That may make it a lot faster on machines that don't have whole gigs
of memory lying around free all the time.

---

By itself, the upgrade causes a revival of the exact same symptoms
described in 833d083a9: release builds crash on startup, because the
JS bundle is nowhere to be found.  On investigation, this is...
exactly a rerun of the same upstream RN issue, in exactly the way
predicted in that commit message:

  Further poking around finds where the old directory name is coming
  from: `react.gradle` hardcodes it.  Fixed upstream in commit
  facebook/react-native@e90319e9f (facebook/react-native#21782)...
  though by hardcoding a different scheme, so that still seems
  vulnerable to breaking on a future plugin version. 🤷

Apparently it did.  The new directory the assets are taken from is
no longer
  app/build/intermediates/merged_assets/release/mergeReleaseAssets/out
but the less redundant
  app/build/intermediates/merged_assets/release/out
So, use that.