webview: Don't assume message-peers have `data-msg-id`.
At present, we set the `data-msg-id` attribute not only on actual
message elements but also on the siblings that come between them:
timerows and recipient headers.  For those not-actual-messages,
the value we use is the message ID of the *next* message.

The original function of doing that was to make the old
visible-messages algorithm work.  As described in commit 7f968e568
which replaced that algorithm, this led to bugs where we would see a
timerow or recipient header in view, use its `data-msg-id` which in
truth belonged to a message that was nowhere near onscreen, and mark
messages as read that the user had never actually seen.

And in fact it was pretty predictable that this would be a source of
bugs -- in general, setting a "message ID" attribute on an element
which doesn't actually correspond to any particular message is a
recipe for confusion and buggy logic.

So, we should avoid adding any logic that relies on a `data-msg-id`
attribute being present on anything other than a message element.
A recent commit added such logic in the form of this loop.  Adjust
it to remove that assumption.