webview: iOS now fires scroll events properly, yay!
More precisely, this problem apparently only affected UIWebView;
it was fixed in WKWebView, which Apple shipped way back in iOS 8
and recommended people move to, and which RN started supporting
in v0.57.  Now that we've switched to WKWebView, this shouldn't
be a problem for us anymore.

See note on MDN here:
  https://developer.mozilla.org/en-US/docs/Web/API/Document/scroll_event
and discussion in the linked issue thread.