msglist css: Move night-mode style rules to their own file.
I've been confused a couple of times by diffs which seemed to add
or change redundant CSS rules; it's been because one of them was
specific to night mode, and it's hard to tell that in our current
code structure.

Similarly, if a diff changes something only in night mode, or if
something applying to night mode shows up in a grep, it's hard to
see that from the immediate context.

So, move the night-mode CSS to its own file.  That way diffs and
grep results alike will make clear what rules apply in general or
only in night mode.

Also replace where we've used the term "dark".  We consistently
call it "night mode" in the UI and the rest of the code, and the
webapp calls it that too.