msglist: Adjust various lengths to multiples of 1/16 rem.
These are lengths we intend to convert into px units, to make them
consistent with the layout in general.  Their current values don't
correspond to integer multiples of a px; so adjust them to values
that will, and that work well.

(Some of the values are taken from Boris's change which will be the
next commit, except converted back to rem; others are my adjustments.)