styles: Change color format from RGB to HSL.
Color format has been changed everywhere, except on all tests for
selectors which find colors for different features of the app. This
is because the selectors fetch colors from the backend, and the
reason why this change is happening (described in related issue)
does not concern colors fetched from the backend.

Colors converted using vscode extension 'Color Picker'.
https://marketplace.visualstudio.com/items?itemName=anseki.vscode-color

This is a followup to #3516.