msglist css: Fix a few straggler `em` uses to `rem`.
These slipped in with 8a1099fcc -- the original version of that
change was written before 6ef229e42 was merged and converted our
`rem` uses to `em`, and when we rebased the former we missed that
change and didn't adjust it accordingly.

See the latter commit for motivation.  Note that this one includes
an example where the units actually differ -- the padding and
margin for `.timestamp` had been affected by the `font-size`
setting on the same elements.