time pills: Allow width to vary with content.
By keeping the horizontal padding to a fixed value rather than let it
get wider for narrower times (like "1:11 PM"), I think this makes the
pills look slightly neater.  And thanks to the preceding series of
commits we aren't relying on the width to fine-tune the position,
so we're now free to change this!