android build: Add comments on compileSdkVersion and friends.
The parent commit wasn't the first time I've gone looking up exactly
what compileSdkVersion means and how it's recommended to set it, and
after a while found this same very helpful blog post.  Hopefully now
it can be the last.

Note that SDK 29 (Android Q) is available in the sense that the
build will cheerfully accept it.  But it's still in beta, so I'm
happy holding off (sticking with 28 for now) until it's released as
"stable".  Added a link to the page where it'll appear when it is.