css: Make alert-words more readable.
The previous iteration of alert words was not readily accessible
for people with color blindness.  As @jmandel reports at #3531, he
wasn't able to see the white text on the bright green background
while in dark mode.

This commit is a quick partial fix for the issue, by reducing the
alpha of the background color in order to make the white text on
it more readable (in dark mode).