user selectors: Add selector for retrieving current user.
From the beginning of the Zulip project we have been using Emails as
user identification; it was a legacy choice. We have been migrating
away from that, and as part of the new way of operations, the newer
endpoints use the users `user_id` instead of the Email
id. getOwnUser serves the purpose of getting the current logged in
user in order to fetch their details, like user_id.