data: Act on ACCOUNT_SWITCH more consistently.
Found with `git grep -L 'case ACCOUNT_SWITCH'` where we weren't.
Now we act on these actions in every reducer but `settings`.

This stuff is still kind of a mess (when we act vs. don't act on
LOGOUT, LOGIN_SUCCESS, DEAD_QUEUE, etc.); but pretty sure this change
makes it only less so.