data: Always discard account-specific data on LOGOUT.
Eep, we were missing a bunch of these.

With few exceptions, we want to do the same thing on either LOGOUT or
ACCOUNT_SWITCH: namely, discard the data in favor of an initial state.

We've been a lot better about getting this right for ACCOUNT_SWITCH,
perhaps because that's a much more everyday operation; so a useful
draft came from just adding a LOGOUT case above ACCOUNT_SWITCH, in
each reducer that lacked one:

  $ perl -i -0pe '
        s/ACCOUNT_SWITCH/LOGOUT, ACCOUNT_SWITCH/;
        s/ case ACCOUNT_SWITCH:/ case LOGOUT: $&/
      ' -- $(
        git grep -L 'case LOGOUT' src/*/*Reducer.js
      ) && \
    tools/fmt

The one manual fixup is that I left `sessionReducer` unchanged,
because the ACCOUNT_SWITCH logic there is more complex.  That logic
doesn't look right, but fixing it is out of scope for this commit.