package.json: Capitulate to `yarn`'s auto-formatting.
This makes it less readable... but not hugely so, as the blank
separator lines were always a poor substitute for the appalling
unavailability of comments.

And it makes it a lot more convenient to use `yarn upgrade`
frequently, with specific arguments to operate on different
subsets of our dependencies at a time.