eslint: Silence the rule import/no-cycle .
This is one of the rules that start making noise upon the
upcoming upgrade.  Generally it'd be nice to fix, but each
instance usually requires thoughtful refactoring to fix;
so, well beyond the scope of what it makes sense to tackle
in the course of an upgrade.  Mark it and move on.