eslint: Disable no-else-return.
This is another rule newly empowered with the upcoming upgrade.
(It must have already been in force, because we have this line
disabling it in one spot... but post-upgrade it complains in
many more places.)

As in this instance, it often forbids perfectly sensible code and
just requires making it worse.  And I don't see much corresponding
benefit: at best it could help people avoid deeply-nested `else`
in favor of the early-return pattern, but it seems an ineffective
way to actually teach that, and it's not like this is catching
something subtle and hard to spot.  Just turn it off.