android: Allow plain-HTTP connections in debug builds.
This is actually a no-op in our current configuration, because this
`android:usesCleartextTraffic` attribute defaults to `true` anyway.
When we update `targetSdkVersion` to 28 (#3563), that default will
change to `false` and this will kick in.  See upstream docs:
  https://developer.android.com/guide/topics/manifest/application-element#usesCleartextTraffic

Making this change now to follow upstream's changes to the template;
see in particular facebook/react-native@3b0681527 .