Upgrade React Native to v0.59, from v0.57.
Fixes #3399!

The bulk of the work of this upgrade went into a large number of
commits previous to this one.  The majority of the last few dozen
commits, after 1c86488ce^, were part of the upgrade; some discussion
in #3561.  An earlier wave of effort focused on getting things working
with the upgrade of Flow to v0.92; see #3450 and especially f8c981028,
and then #3453 and #3442.  Some other early discussion is at #3422.

This commit itself was done mainly by running:

  $ tools/upgrade rn v0.59.10

Then, that tool needs a bit of an update for changes upstream.
Because I'm feeling pressed to get this upgrade out the door (and
to deal with the various separate trouble on iOS), I just took
care of the other relevant packages by hand: updated `@babel/core`
and `metro-react-native-babel-preset` in package.json according to
the diff seen in `rn-diff-purge`, then reran `yarn`.