flow: Drop react-native/flow-github directory.
This follows facebook/react-native@5d06c7495 , part of the upgrade
to RN v0.59.

Even while we're still on v0.57, there's not much in that directory --
just a `metro.js` that's a bunch of `any` stubs -- and Flow runs just
fine with it gone from `libs`.  So, we can go ahead and drop it now.