android build: Target 64-bit architectures.
This fixes #3323, meeting the Google Play 64-bit requirement that
comes into effect next week, 2019-08-01 -- in other words, it makes it
possible for us to continue uploading new releases after that date.

Aside from ticking a box, it's expected to improve performance on
devices with 64-bit CPUs -- which is most new devices of the last few
years, and 85% of all our installs on active devices as reported by
the Play Console.  (In "Release management > Device catalog",
filtering "ABI" to `arm64-v8a` or `x86_64`.)

The main work to make this possible happened in RN upstream; we pulled
it in with our upgrade to RN v0.59, #3399.  This is one last fragment
of the diff in upstream's template app between those versions,
enabling 64-bit versions in our build config.

One unfortunate regression this causes: we're still distributing a
single APK for all architectures, and so adding 64-bit architectures
makes it a lot bigger.  At a quick estimate from comparing
`yarn build:android-nokeys` before and after, we go from 13MB to 22MB.

The Play Store has had a solution to that for a while now, called
"Android App Bundle".  We should switch to that.  We're already
tracking that task as #3547, and this change increases its priority.