travis: Stay on working Trusty, while "xenial" is really Precise.
Fixes #3565.

Travis CI is migrating their default build environments from
14.04 Trusty to 16.04 Xenial.  That's more than reasonable --
Trusty went EOL a few months ago.

But for Android they don't yet have a Xenial image.  And when you
request a Xenial image for Android... what you actually get turns
out to be a *12.04 Precise* image.  Which went EOL over two years
ago, and which naturally things don't always run on.

We don't specify an OS release (with `dist`), so we get the default;
and they've been gradually rolling out the new default to different
users, and a week ago it was apparently our turn.  Since then we've
had the supposed `xenial` which is actually a Precise image, and
the build has failed because the `flow` binary expects a newer glibc
than that ancient image has.

Fix it by explicitly sticking with Trusty, as the least-ancient
option available.

More details in the issue thread #3565.  I also just reported the
issue to Travis on their forum:
  https://travis-ci.community/t/android-builds-default-to-precise-which-is-2-years-eol/4352