ios: Drop iOS 9 support; require iOS 10.3+.
This is the oldest version I can actually get a simulator for in the
current version of Xcode, making it the oldest that we can
realistically actually keep the app working properly on.

Already by April this year when I last looked, <0.4% of our iOS users
were on iOS 9 (and 3.1% on <=10), so this will affect very few users.

Also allow graceful degradation on iOS 10 and 11.  In the same April
figures, versions <=11 were 10% of iOS users, so graceful degradation
at that level is fine.