getAggregatedPresence: Simplify and tighten comparator logic.
There's no need to say `if (foo) { return true; }` when the next step
will always be to return false; just return foo in the first place.

(In general the rewrite would be to return `!!foo`, but in this case
the values are manifest booleans already.)

Also mitigate the verbosity of `case` with a bit of formatting.