internal links: Pull out parsing of stream and topic operands.
This creates a good home for the complexity we need to add to
the `stream` case to address #2760.

These two implementations are identical, but that's sort of an
accident -- in fact, a bug, namely #2760, as the real format of
stream operands has been more complex than this for some time.