compose menu: Defer import of react-native-document-picker.
This library prints an obnoxious warning right at import time,
if its native module isn't present.  (Actually worse: at import
time it calls `setTimeout` with a callback to print the warning.)

Running under Jest, the native module is of course not present.
As a result, we were getting a flood of this warning on every
`tools/test` run.

Work around the obnoxious warning by deferring import to inside
the handler where we're about to actually invoke the library.