yarn: Take all stable updates to dependencies.
Done with:

  $ yarn upgrade
  $ yarn update-libdefs

and then running `tools/test --full`, and playing a bit with a
debug build, to confirm all seems well.

As before, `yarn upgrade` generates some warnings from various
indirect dependencies, all some form of deprecation notice.

The list is the same as in commit d63b8db11, except that the warning
about `jest > jest-cli > prompts > kleur@2.0.2` is now gone.
Consulting the history, it looks like that was resolved in 6db8bd67f
where we upgraded `jest` and `jest-cli` to latest, pulling in a newer
`prompts` (was ^0.1.9 -> 0.1.14, became ^2.0.1 -> 2.1.0) that in turn
had upgraded to the newer `kleur`.