nav: Update some react-navigation usage for deprecations in v2.
This follows up on the parent commit upgrading to react-navigation v2,
by updating our use of it to follow two changes in the API.

Before this change, everything works but we get several warnings.
This change resolves two of the warnings:

    console.warn node_modules/react-navigation/src/react-navigation.js:63
      TabNavigator is deprecated. Please use the createBottomTabNavigator
      or createMaterialTopTabNavigator instead.
    console.warn node_modules/react-navigation/src/react-navigation.js:32
      The StackNavigator function name is deprecated, please use
      createStackNavigator instead

After this change, we still get three warnings, including another copy
of the TabNavigator one.  Resolving those involves pulling in another
package; we'll do that in the next commit.