nav: Cut misleading `subscriptions` route and screen.
This nav action creator is never called... which means this route
is never navigated to... and that was the only reference to this
component.  So they're all dead code and only cause confusion;
cut them out.

Good thing, too -- among other things, the name said "all streams"
but then the component name said "subscriptions", and it actually
showed the latter.