api.queueMarkAsRead: Set timeout to follow up on a debounced request.
If there are multiple requests to mark messages as read with interval
less than 2s, and if there is no further request soon after the 2s
mark, then the later requests remain stuck in the queue and aren't
sent to the server.

One symptom of this would be that after reaching the end of the
message list, the unread banner might still be visible with some
unread count.

So now set timeout to send read-message flag to server.

Fixes: #3509