auth: Update Google-auth URL, following Zulip server change.
Fixes #3594.

The Zulip server recently (zulip/zulip@bf14a0af4, in July) refactored
its support for Google auth, to use the same third-party library we're
using for GitHub auth and Azure AD auth.

One side effect is that the URL the Zulip server provides for starting
the Google auth flow has changed.  Confusingly, the old URL now almost
seems to work -- it gives a generic Zulip login page -- but then if
you complete the login you wind up in the Zulip webapp, still inside
the browser view which we created for the login flow.

Handle this by updating the URL to the new one.

This is a partial fix, as a client-side fix for a breaking server-side
change always is: it will resolve the issue for users who get the new
version of the app, but not for users on all the already-deployed
versions of the app.  For those, we'll need to fix the server.