auth: Revert to old Google-auth URL, for old servers' sake.
This reverts f466e3bc1.

We've now fixed #3594 on the server side (as zulip/zulip#13081),
so that the server accepts both the old and the new URL paths.
That update has been deployed on zulipchat.com...  and meanwhile,
all servers running a released version of Zulip, or a version
older than last month's zulip/zulip@bf14a0af4, accept only the
old URL path.

So, switch back to the old path.