android build: Use our patched react-native-webview to cut noise in log.
I've just made this "fork" repo, and a `zulip` branch in it, in order
to make a fix for a noisy bit of logging that was showing up even in
`tools/test android`, where we suppress non-error logs.

This commit zulip/react-native-webview#49aca8206 consists of a single
small patch atop v5.12.1, which we had been using.

I sent the fix as a PR upstream:
  react-native-community/react-native-webview#844
which was promptly merged (thanks, upstream maintainers!), and
even promptly released, as 7.0.3.  We're still on 5.x until we
deal with the RN upgrade to v0.60 (#3548), so for now we stick
the patch in this small fork-branch of our own.

(This is one of those times I really wish we had comments in
package.json. :-/  The key information in this commit message
really ought to be in a 1- or 2-line comment in the file.)