docs/crunchy-shell: Linkify commit IDs.
Sadly GitHub doesn't automatically do this in Markdown files found in
the repo, though it does in commit messages, issue/PR threads, etc.
(I guess I probably wouldn't always be happy if it did so
unconditionally, but it'd sure be nice to be able to opt into it.)

Did this mainly with a quick ad-hoc editor macro.