android notif: Cut disused bits for saying "(2 messages)".
The grandparent commit cut the logic that was using this.

Which is a good thing, too, because this was always a buggy way to
have implemented this logic -- it would have broken as soon as we'd
started seriously using this `plurals` facility for translation, which
is what it's actually for.

In fact the upstream docs warn quite specifically against this:

  It would be a mistake to use quantity strings to implement something
  like Gmail's "Inbox" versus "Inbox (12)" when there are unread
  messages, for example.  It might seem convenient to use quantity
  strings instead of an `if` statement, but it's important to note
  that some languages (such as Chinese) don't make these grammatical
  distinctions at all, so you'll always get the `other` string.

  https://developer.android.com/guide/topics/resources/string-resource.html#Plurals

For an example of using this `plurals` feature as intended, on the
other hand, see the `numConversations` item just recently added.