dependencies: Use ordinary NPM version of r-n-document-picker.
The existing unusual git:// reference for `react-native-document-picker`
was causing `flow-typed` to generate a stub file it thereafter couldn't
read.

Replace this (which seems to have effectively been 3.2.0) with an
ordinary ref to ^3.2.4.

We originally added this in d9bbb5e5f, based on someone's installation
guide upstream.  Looking back at that guide, there isn't an obvious
reason why a Git reference was preferred to an NPM one.