eventMiddleware: Set event.message.flags to [] even if absent on event.
The conditional described by the comment "move `flags` key from
`event` to `event.message` for consistency" was not defaulting
`event.message.flags` to an empty array in cases where `flags` is
absent on `event`. Defaulting to an empty array will give more
resistance to unpredictable input, in a crunchy shell pattern:

https://github.com/zulip/zulip-mobile/blob/master/docs/architecture/crunchy-shell.md

This commit replaces a test from 25e97ef8 (fixed #3046) that ensured
the opposite (!): that `event.message` was not mutated in the case
where `event.message.flags` was not set. I didn't see an obvious
reason for this test, and an explanation wasn't given in the commit
message. If possible, I do think it's important to have a dependable
empty value for event.message.flags, so we don't have to constantly
check internally if it's defined.