logging: Migrate all `logErrorRemotely` calls to new `logging.*` API.
Two reasons to get "remotely" out of the function's name:

 * It was already inaccurate -- this function was just as much for
   logging locally as remotely.

 * If accurate, it wouldn't belong in the API.  The choice that
   makes sense to express at individual error sites is between
   message severity levels (error, warning, etc.), while specific
   reporting routes (local logs, Sentry, etc.) are a policy choice
   to make centrally per severity level.

The `msg` parameter goes away because there isn't an obvious way to
shoehorn it in for Sentry reporting, and because it was never that
helpful: we were already dropping it on the floor when reporting to
Sentry, and anyway there's the stack trace to tell the story.

The call sites become `logging.error` or `logging.warn` calls
according to their severity, as described in those functions' jsdoc.
Where there's a useful Error object, we pass that; else we make up
a string.