android notif [nfc]: Factor out logic to distill lifecycle status.
This makes `notifyReact` quite a bit shorter by squeezing out most of
the logic (and comments) for inspecting the RN lifecycle status into
an enum plus a general, pure function.

This better isolates the impure, application-specific logic.  It lets
`notifyReact` focus on that part, and in particular it lets us
deduplicate the `initialNotification` and `launchMainActivity` lines.