webview js: Manually absolutize <img> tags' sources
We've hitherto relied on the WebView to make relative image links
absolute. However, it does this using our supplied `baseUrl`, which is
about to change.

Rewrite the `appendAuthToImages` function to explicitly rewrite image
source URLs based on `auth.realm`, rather than leaving the WebView to
do so implicitly based on `document.location`. Rename the function
appropriately.

In the WebView's present configuration, this introduces no functional
changes (as witnessed by the test suite).