android: Bump minSdkVersion to 21, aka Android 5 Lollipop.
This begins with a correction to what we say about browser versions
in the WebView!  I'd read the documentation last year to say that
starting with Android 4.4 KitKat, the WebView's browser updated as
an APK independent of the OS... but in fact that only began with
Android 5.

Beyond that correction, in this commit we
 * update the authoritative statements of our minimum version,
   for both humans and metadata-reading computers;
 * add an entry to our historical log of when and why we
   dropped older versions.

This change also permits a few simplifications in the actual code,
but we leave those for separate followup commits.