AuthScreen [nfc]: Drive methods with genuine plain old data.
This lets us stop pulling strings out of a table and invoking the
functions that have their names.

It also lets us treat the four OAuth methods more generically --
the thing to do for each of them differs only by URL, so make the
URLs data instead of embedding them in separate function definitions.

For the type of `action` we could do something more elaborate like

    {| type: 'dev' |} | ... | {| type: 'oauth', url: string |}

and that'd be the right thing if this were an API spread widely.
Because it's entirely internal to this module, though, that
doesn't seem necessary.