webAuth [nfc]: Move extractApiKey inside this module.
Its job is wholly specific to this protocol.

Also drop a meaningless unit test that just recites the
implementation.  A proper unit test for this would actually
provide the input and expected output... and in an upcoming
commit, we'll do exactly that as part of testing the code that
this is a tiny helper function for.